Rename "web3" directory to "web3Data"
