+++ src/blockchainData/useCurrentTimestamp.ts
@@ -1,5 +1,5 @@
 import { useEffect, useState } from "react";
-import { useWeb3 } from "../web3";
+import { useWeb3 } from "../web3Data";
 
 const useCurrentTimestamp = (blockNumber: number | undefined) => {
   const [timestamp, setTimestamp] = useState<number>(
