+++ src/daoData/useAccessControlContract.ts
@@ -1,6 +1,6 @@
 import { useState, useEffect } from 'react';
 import { AccessControl, AccessControl__factory } from '../typechain-types';
-import { useWeb3 } from '../web3';
+import { useWeb3 } from '../web3Data';
 
 const useAccessControlContract = (accessControlAddress: string | undefined) => {
   const [accessControlContract, setAccessControlContract] = useState<AccessControl>();
