+++ src/daoData/useDAOContract.ts
@@ -1,7 +1,7 @@
 import { useState, useEffect } from 'react';
 
 import { DAO, DAO__factory } from '../typechain-types';
-import { useWeb3 } from '../web3';
+import { useWeb3 } from '../web3Data';
 
 const useDAOContract = (address: string | undefined) => {
   const [dao, setDAO] = useState<DAO>();
