+++ src/daoData/useExecuteTransaction.ts
@@ -1,6 +1,6 @@
 import { useCallback } from 'react'
-import { useTransaction } from '../web3/transactions';
-import { useWeb3 } from '../web3';
+import { useTransaction } from '../web3Data/transactions';
+import { useWeb3 } from '../web3Data';
 import { GovernorModule, GovernorModule__factory } from '../typechain-types';
 import { useDAOData } from './index';
 import { ProposalData } from "../daoData/useProposals";
