+++ src/hooks/useENSName.ts
@@ -1,6 +1,6 @@
 import { useState, useEffect } from 'react';
 
-import { useWeb3 } from '../web3';
+import { useWeb3 } from '../web3Data';
 
 const useENSName = (account: string | undefined) => {
   const [{ provider }] = useWeb3();
